+++ src/main/java/ru/kolaer/permit/service/EmployeePageService.java
@@ -10,7 +10,7 @@ public interface EmployeePageService extends BasePageService<EmployeeEntity> {
     EmployeeEntity getByPersonnelNumber(Integer personnelNumber);
     EmployeeEntity getByUsername(String username);
 
-    Integer getIdByPersonnelNumber(Integer personnelNumber);
+    Long getIdByPersonnelNumber(Long personnelNumber);
 
 
 }
